Move crypto stuff to CryptoUtils
